Redid cube class, added some movement
Reworked the whole cube class to make it easier to turn, I also added a UI option for solving method which will be used later once the cube has full turning capabilities, and added some basic turning logic which isn't fully functional. Also added code for a scramble sequence generator which is currently not used but will be.