Bug Fixes + New scramble sequence generation
Fixed some performance bugs when making a turn on the cube, large cubes (above 15x15) seem to still suffer in perfomance.
Reworked the scramble generator and move sequence reader. In the scramble I changed it to generate by row number to fix the probability of turning each possible move. The turn function was changed by calling it with an array of objects instead of a string. The object includes the move sequence, row, and direction. This fixed a bug with turning a row greater than 9 as if it read '10' for example, it would only look at the first index and turn row #1. Also removed piece.js as it is not being used.