Created files with blank functions for each solving algorithm.
Created the template for functions on the beginner method and created files for each of the solving algorithms.
Reverse function also does not work if you've scrambled it from a non solved state. To fix this, I haven't decided whether or not to append the turning sequences to get the full history of moves, or to reset the cube back to a solved state when you go to scramble it again.
Also added a favicon.